+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
 - feat: add round_idx to get_performance() (#192)
+
+### Changed
+
+- feat: Add algo named inputs and outputs for existing task in remote mode (#191)
 - feat: log level set to warning for connect-tools (#185)
 
 ## [0.21.0](<https://github.com/owkin/substra/releases/tag/0.21.0>) - 2022-05-16